feat: Paypal invoicing.
Added the ability to generate invoices via PayPal for non-shield
deliverables. PayPal webhooks will then allow these invoices to be moved
forward upon payment/cancellation/refund, and the amounts on the
invoices will be synced to their PayPal counterparts.

Some preliminary work was made in this commit to support currencies
other than US Dollars: The backend line calculations can now handle
non-US dollars consistently (or should be able to) and invoices can now
be annotated with a specific currency in the DB. As of yet, there's no
interface for using this, but it is used by the calculation functions.