feat: Enable GA for third party cookie consent
On the advice of a marketing consultant who uses it extensively, we're
adding Google Analytics in for those who have consented to third
party cookies.

We'll be watching how they use Google Analytics to determine if we
can go without it and have them just use Matomo, which we're already
paying for and prefer.